remove non-parallel semantic sectioning
