+++ dsrag/dsparse/tests/unit/test_semantic_sectioning.py
@@ -2,82 +2,269 @@ import os
 import sys
 import unittest
 import json
+from unittest.mock import patch, Mock, MagicMock
 
 sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '../../..')))
 sys.path.append(os.path.join(os.path.dirname(__file__), '..', '..'))
 from dsparse.sectioning_and_chunking.semantic_sectioning import (
-    get_document_with_lines, 
-    validate_and_fix_sections, 
-    split_long_line, 
-    elements_to_lines, 
-    str_to_lines, 
-    pages_to_lines, 
-    DocumentSection, 
+    get_document_text_for_window,
+    validate_and_fix_window_sections,
+    split_long_line,
+    elements_to_lines,
+    str_to_lines,
+    pages_to_lines,
+    DocumentSection,
     get_sections_text,
+    create_document_windows,
+    merge_sections_across_windows,
+    validate_and_fix_global_sections,
+    get_sections,
+    get_sections_from_str_parallel,
 )
 
 class TestSemanticSectioning(unittest.TestCase):
-    def test__get_document_with_lines(self):
-        # Test that the document with line numbers is returned correctly
-        document_lines_path = os.path.abspath(os.path.join(os.path.dirname(__file__), '../data/document_lines.json'))
-        with open(document_lines_path, 'r') as f:
-            document_lines = json.load(f)
-
-        print (len(document_lines))
-
-        document_with_line_numbers, end_line = get_document_with_lines(document_lines, start_line=0, max_characters=20000)
-        split_lines = document_with_line_numbers.split("\n")
-
-        self.assertEqual(split_lines[0], "[0] McKinsey")
-        self.assertEqual(end_line, 203)
+    def setUp(self):
+        # Create sample document lines for testing
+        self.document_lines = [{"content": f"Line {i}"} for i in range(100)]
+        
+        # Sample document string
+        self.sample_document = "\n".join([f"Line {i}" for i in range(100)])
+        
+        # Data path for loading test data
+        self.data_path = os.path.abspath(os.path.join(os.path.dirname(__file__), '../data'))
+        
+    def test_document_text_for_window(self):
+        """Test that window text is generated correctly with proper line numbers"""
+        window_text = get_document_text_for_window(self.document_lines, 10, 20)
+        
+        # Verify the first and last line have correct line numbers
+        lines = window_text.strip().split("\n")
+        self.assertTrue(lines[0].startswith("[10]"))
+        self.assertTrue(lines[-1].startswith("[20]"))
+        
+        # Verify correct number of lines
+        self.assertEqual(len(lines), 11)  # 10-20 inclusive = 11 lines
+        
+        # Test boundary conditions
+        window_text = get_document_text_for_window(self.document_lines, 90, 110)
+        lines = window_text.strip().split("\n")
+        self.assertTrue(lines[0].startswith("[90]"))
+        # Should be truncated to document length
+        self.assertTrue(lines[-1].startswith("[99]"))
     
-    def test_validate_and_fix_sections(self):
+    def test_validate_and_fix_window_sections(self):
+        """Test that window section validation works correctly"""
         # Test with valid sections
-        sections = [
-            DocumentSection(title="Section 1", start_index=0),
-            DocumentSection(title="Section 2", start_index=5),
-            DocumentSection(title="Section 3", start_index=10)
+        window_sections = [
+            DocumentSection(title="Window Section 1", start_index=10),
+            DocumentSection(title="Window Section 2", start_index=15),
+            DocumentSection(title="Window Section 3", start_index=18)
         ]
-        fixed = validate_and_fix_sections(sections, document_length=15)
-        self.assertEqual(len(fixed), 3)
-        self.assertEqual([s.start_index for s in fixed], [0, 5, 10])
-
-        # Test with out of order sections
-        sections = [
-            DocumentSection(title="Section 2", start_index=5),
-            DocumentSection(title="Section 1", start_index=0),
-            DocumentSection(title="Section 3", start_index=10)
+        
+        fixed = validate_and_fix_window_sections(
+            window_sections, 
+            window_start_line=10, 
+            window_end_line=20,
+            document_length=100
+        )
+        
+        # First section should start at window start
+        self.assertEqual(fixed[0].start_index, 10)
+        
+        # Test with out-of-order sections
+        window_sections = [
+            DocumentSection(title="Window Section 2", start_index=15),
+            DocumentSection(title="Window Section 1", start_index=10),
+            DocumentSection(title="Window Section 3", start_index=18)
         ]
-        fixed = validate_and_fix_sections(sections, document_length=15)
-        self.assertEqual([s.start_index for s in fixed], [0, 5, 10])
-        self.assertEqual(fixed[0].title, "Section 1")
-        self.assertEqual(fixed[1].title, "Section 2")
-        self.assertEqual(fixed[2].title, "Section 3")
-
-        # Test with duplicate start indices
-        sections = [
-            DocumentSection(title="Section 1", start_index=0),
-            DocumentSection(title="Section 2", start_index=5),
-            DocumentSection(title="Section 3", start_index=5)
+        
+        fixed = validate_and_fix_window_sections(
+            window_sections, 
+            window_start_line=10, 
+            window_end_line=20,
+            document_length=100
+        )
+        
+        # Sections should be ordered
+        self.assertEqual(fixed[0].title, "Window Section 1")
+        self.assertEqual(fixed[1].title, "Window Section 2")
+        self.assertEqual(fixed[2].title, "Window Section 3")
+        
+        # Test with section beyond window
+        window_sections = [
+            DocumentSection(title="Window Section 1", start_index=10),
+            DocumentSection(title="Window Section 2", start_index=25)  # Beyond window end
         ]
-        fixed = validate_and_fix_sections(sections, document_length=15)
-        self.assertEqual(len(fixed), 2)
-        self.assertEqual([s.start_index for s in fixed], [0, 5])
-        self.assertEqual(fixed[0].title, "Section 1")
-        self.assertEqual(fixed[1].title, "Section 2")
-
+        
+        fixed = validate_and_fix_window_sections(
+            window_sections, 
+            window_start_line=10, 
+            window_end_line=20,
+            document_length=100
+        )
+        
+        # Section beyond window should be removed
+        self.assertEqual(len(fixed), 1)
+        
+        # Test with empty sections list
+        fixed = validate_and_fix_window_sections(
+            [], 
+            window_start_line=10, 
+            window_end_line=20,
+            document_length=100
+        )
+        
+        # Should create a default section
+        self.assertEqual(len(fixed), 1)
+        self.assertEqual(fixed[0].start_index, 10)
+    
+    def test_create_document_windows(self):
+        """Test window creation works correctly"""
+        # Create a test document with varying line lengths
+        lines_with_varying_lengths = []
+        for i in range(100):
+            # Some lines are long, some are short
+            content = f"Line {i}" + " extra content" * (i % 5)
+            lines_with_varying_lengths.append({"content": content})
+        
+        # Test with small window size to force multiple windows
+        windows = create_document_windows(lines_with_varying_lengths, max_characters_per_window=500)
+        
+        # Should have multiple windows
+        self.assertTrue(len(windows) > 1)
+        
+        # Windows should cover the entire document
+        self.assertEqual(windows[0][0], 0)  # First window starts at beginning
+        self.assertEqual(windows[-1][1], 99)  # Last window ends at end
+        
+        # Windows should not overlap
+        for i in range(len(windows) - 1):
+            self.assertEqual(windows[i][1] + 1, windows[i+1][0], 
+                           f"Window {i} and {i+1} overlap or have gap")
+        
+        # Test with empty document
+        empty_windows = create_document_windows([], max_characters_per_window=500)
+        self.assertEqual(len(empty_windows), 0)
+    
+    def test_merge_sections_across_windows(self):
+        """Test section merging across windows"""
+        # Create sections from 3 windows
+        window1_sections = [
+            DocumentSection(title="Window 1 Section 1", start_index=0),
+            DocumentSection(title="Window 1 Section 2", start_index=5)
+        ]
+        
+        window2_sections = [
+            DocumentSection(title="Window 2 Section 1", start_index=10),
+            DocumentSection(title="Window 2 Section 2", start_index=15)
+        ]
+        
+        window3_sections = [
+            DocumentSection(title="Window 3 Section 1", start_index=20),
+            DocumentSection(title="Window 3 Section 2", start_index=25)
+        ]
+        
+        all_window_sections = [window1_sections, window2_sections, window3_sections]
+        
+        merged_sections = merge_sections_across_windows(all_window_sections)
+        
+        # Sections after merging: W1S1, merged W1S2/W2S1, W2S2, merged W2S2/W3S1, W3S2
+        # The number of sections should be 5 since we have 3 windows with 2 sections each
+        # and each adjacent window pair gets 1 section merged, so: 6 - 1 = 5 sections
+        self.assertEqual(len(merged_sections), 5)
+        
+        # First section should be preserved from window 1
+        self.assertEqual(merged_sections[0].title, "Window 1 Section 1")
+        
+        # Last section from window 1 should be merged with first section from window 2
+        self.assertTrue("Window 1 Section 2 / Window 2 Section 1" in merged_sections[1].title)
+        
+        # Test with empty window sections list
+        merged_empty = merge_sections_across_windows([])
+        self.assertEqual(len(merged_empty), 0)
+        
+        # Test with single window
+        merged_single = merge_sections_across_windows([window1_sections])
+        self.assertEqual(len(merged_single), 2)
+        self.assertEqual(merged_single[0].title, "Window 1 Section 1")
+    
+    def test_validate_and_fix_global_sections(self):
+        """Test global section validation"""
+        # Create sections for testing
+        sections = [
+            DocumentSection(title="Section 1", start_index=10),  # Should be adjusted to 0
+            DocumentSection(title="Section 2", start_index=20),
+            DocumentSection(title="Section 3", start_index=30)
+        ]
+        
+        fixed = validate_and_fix_global_sections(sections, document_length=100)
+        
+        # First section should start at 0
+        self.assertEqual(fixed[0].start_index, 0)
+        
+        # Test with empty sections list
+        fixed_empty = validate_and_fix_global_sections([], document_length=100)
+        self.assertEqual(len(fixed_empty), 1)
+        self.assertEqual(fixed_empty[0].start_index, 0)
+        
+        # Test with sections beyond document bounds
+        beyond_bounds = [
+            DocumentSection(title="Section 1", start_index=0),
+            DocumentSection(title="Section 2", start_index=110)  # Beyond document length
+        ]
+        
+        fixed_bounds = validate_and_fix_global_sections(beyond_bounds, document_length=100)
+        self.assertEqual(len(fixed_bounds), 1)  # Section beyond bounds should be removed
+    
+    @patch('dsparse.sectioning_and_chunking.semantic_sectioning.get_structured_document_for_window')
+    def test_parallel_get_sections(self, mock_get_structured_doc):
+        """Test parallel processing of document sections"""
+        # Mock the LLM call to return predetermined sections
+        mock_response = Mock()
+        mock_response.sections = [
+            DocumentSection(title="Test Section 1", start_index=0),
+            DocumentSection(title="Test Section 2", start_index=5)
+        ]
+        mock_get_structured_doc.return_value = mock_response
+        
+        # Process document with mocked LLM calls
+        sections = get_sections(
+            document_lines=self.document_lines,
+            max_characters_per_window=500,
+            llm_provider="openai",
+            model="gpt-4o-mini",
+            language="en",
+            max_concurrent_llm_calls=2
+        )
+        
+        # Should have produced sections
+        self.assertTrue(len(sections) > 0)
+        
+        # Each section should have title and content
+        for section in sections:
+            self.assertIn("title", section)
+            self.assertIn("content", section)
+            self.assertIn("start", section)
+            self.assertIn("end", section)
+    
     def test_split_long_line(self):
-        # Test line that doesn't need splitting
+        """Test splitting long lines"""
+        # Short line doesn't need splitting
         short_line = "This is a short line"
         self.assertEqual(split_long_line(short_line, max_line_length=50), [short_line])
-
-        # Test line that needs splitting
+        
+        # Long line needs splitting
         long_line = "This is a very long line that needs to be split into multiple lines because it exceeds the maximum length"
         split_lines = split_long_line(long_line, max_line_length=20)
+        
+        # Each line should be under max length
         self.assertTrue(all(len(line) <= 20 for line in split_lines))
+        
+        # Combined content should match original
         self.assertEqual(" ".join(split_lines), long_line)
-
+    
     def test_elements_to_lines(self):
+        """Test converting elements to lines"""
         elements = [
             {"type": "Title", "content": "Test Document"},
             {"type": "NarrativeText", "content": "This is a test.\nThis is another line."},
@@ -87,216 +274,68 @@ class TestSemanticSectioning(unittest.TestCase):
         
         # Test with no exclusions
         lines = elements_to_lines(elements, exclude_elements=[], visual_elements=["Image"])
+        
+        # Should have visual and non-visual elements
         self.assertTrue(any(line["is_visual"] for line in lines))
         self.assertTrue(any(not line["is_visual"] for line in lines))
-
+        
         # Test with exclusions
         lines = elements_to_lines(elements, exclude_elements=["Table"], visual_elements=["Image"])
         self.assertFalse(any(line["element_type"] == "Table" for line in lines))
-
+    
     def test_str_to_lines(self):
+        """Test converting string to lines"""
         text = "Line 1\nLine 2\nA very long line that should be split into multiple lines"
         lines = str_to_lines(text, max_line_length=20)
+        
+        # Each line should be under max length
         self.assertTrue(all(len(line["content"]) <= 20 for line in lines))
+        
+        # All lines should be narrative text
         self.assertTrue(all(line["element_type"] == "NarrativeText" for line in lines))
-
+    
     def test_pages_to_lines(self):
+        """Test converting pages to lines"""
         pages = [
             "Page 1 content\nMore content",
             "Page 2 content\nEven more content"
         ]
         lines = pages_to_lines(pages)
+        
+        # Lines should have page numbers
         self.assertEqual(lines[0]["page_number"], 1)
         self.assertEqual(lines[2]["page_number"], 2)
-
-    def test_validate_and_fix_sections_bounds(self):
-        # Test with section starting beyond document length
-        sections = [
-            DocumentSection(title="Section 1", start_index=0),
-            DocumentSection(title="Section 2", start_index=20)  # Beyond doc length
-        ]
-        fixed = validate_and_fix_sections(sections, document_length=15)
-        self.assertEqual(len(fixed), 1)
-        self.assertEqual(fixed[0].start_index, 0)
-
-        # Test with all sections beyond document length
-        sections = [
-            DocumentSection(title="Section 1", start_index=20),
-            DocumentSection(title="Section 2", start_index=25)
-        ]
-        fixed = validate_and_fix_sections(sections, document_length=15)
-        self.assertEqual(len(fixed), 1)
-        self.assertEqual(fixed[0].start_index, 0)
-
-        # Test with section that are out of order with one section beyond document length
-        sections = [
-            DocumentSection(title="Section 1", start_index=0),
-            DocumentSection(title="Section 3", start_index=20),
-            DocumentSection(title="Section 2", start_index=10)
-        ]
-        fixed = validate_and_fix_sections(sections, document_length=15)
-        self.assertEqual(len(fixed), 2)
-        self.assertEqual(fixed[0].start_index, 0)
-        self.assertEqual(fixed[1].start_index, 10)
-        self.assertEqual(fixed[0].title, "Section 1")
-        self.assertEqual(fixed[1].title, "Section 2")
-
-    def test_get_sections_text_bounds(self):
-        document_lines = [
-            {"content": f"Line {i}"} for i in range(10)
-        ]
+    
+    @patch('dsparse.sectioning_and_chunking.semantic_sectioning.parallel_get_sections')
+    def test_get_sections_from_str_parallel(self, mock_parallel_get_sections):
+        """Test the main entry point function with string input"""
+        # Mock the parallel processing function
+        mock_parallel_get_sections.return_value = [{"title": "Test", "content": "Content", "start": 0, "end": 10}]
         
-        # Test with section ending beyond document length
-        sections = [
-            DocumentSection(title="Section 1", start_index=0),
-            DocumentSection(title="Section 2", start_index=8)
-        ]
-        result = get_sections_text(sections, document_lines)
-        self.assertEqual(len(result), 2)
-        self.assertEqual(result[-1]["end"], 9)
-
-    def test_get_sections_iterative_processing(self):
-        """Test the iterative document processing to ensure sections align properly"""
-        # Create a document that needs multiple iterations
-        document_lines = [{"content": f"Line {i}"} for i in range(300)]
+        # Use the function with a string input
+        sections, document_lines = get_sections_from_str_parallel(
+            document=self.sample_document,
+            max_characters_per_window=500,
+            semantic_sectioning_config={"use_semantic_sectioning": True},
+            chunking_config={}
+        )
         
-        # Simulate the iterative processing that happens in get_sections
-        all_sections = []
-        start_line = 0
-        max_iterations = 5
+        # Should have at least one section
+        self.assertTrue(len(sections) > 0)
         
-        for _ in range(max_iterations):
-            doc_str, end_line = get_document_with_lines(document_lines, start_line, max_characters=100)
-            
-            # Simulate what the LLM would return
-            new_sections = [
-                DocumentSection(title=f"Section {i}", start_index=start_line + i*20)
-                for i in range(3)
-            ]
-            
-            # Validate and fix the sections
-            fixed_sections = validate_and_fix_sections(new_sections, len(document_lines))
-            
-            if not fixed_sections:
-                start_line = end_line + 1
-                continue
-            
-            all_sections.extend(fixed_sections)
-            
-            if end_line >= len(document_lines) - 1:
-                break
-            
-            if len(fixed_sections) > 1:
-                # Start from the beginning of the last section to ensure proper overlap handling
-                start_line = all_sections[-1].start_index
-                all_sections.pop()
-            else:
-                start_line = end_line + 1
-            
-            if start_line >= len(document_lines):
-                break
+        # Document lines should be populated
+        self.assertTrue(len(document_lines) > 0)
         
-        # Get the final sections
-        final_sections = get_sections_text(all_sections, document_lines)
+        # Test with semantic sectioning disabled
+        sections, document_lines = get_sections_from_str_parallel(
+            document=self.sample_document,
+            semantic_sectioning_config={"use_semantic_sectioning": False}
+        )
         
-        # Verify no sections extend beyond document bounds
-        self.assertLess(final_sections[-1]["end"], len(document_lines))
-        
-        # Verify sections are continuous (no gaps)
-        for i in range(len(final_sections) - 1):
-            self.assertEqual(final_sections[i]["end"] + 1, final_sections[i + 1]["start"], 
-                            f"Gap between sections {i} and {i+1}")
-        
-        # Verify all sections are within bounds
-        for section in final_sections:
-            self.assertGreaterEqual(section["start"], 0)
-            self.assertLess(section["end"], len(document_lines))
-            self.assertLess(section["start"], section["end"])
-
-    def test_get_sections_overlap_handling(self):
-        """Test that section overlaps are handled correctly"""
-        document_lines = [{"content": f"Line {i}"} for i in range(100)]
-        
-        # Create overlapping sections
-        sections = [
-            DocumentSection(title="Section 1", start_index=0),
-            DocumentSection(title="Section 2", start_index=40),  # Overlaps with Section 3
-            DocumentSection(title="Section 3", start_index=30)   # Overlaps with Section 2
-        ]
-        
-        fixed = validate_and_fix_sections(sections, document_length=len(document_lines))
-        result = get_sections_text(fixed, document_lines)
-        
-        # Verify sections are properly ordered and don't overlap
-        for i in range(len(result) - 1):
-            self.assertLess(result[i]["end"], result[i + 1]["start"], 
-                           f"Sections {i} and {i+1} overlap")
-
-    def test_get_document_with_lines_bounds(self):
-        """Test that get_document_with_lines respects document bounds"""
-        document_lines = [{"content": f"Line {i}"} for i in range(50)]
-        
-        # Test starting from middle
-        doc_str, end_line = get_document_with_lines(document_lines, start_line=25, max_characters=1000)
-        self.assertLess(end_line, len(document_lines))
-        
-        # Test with start_line near end of document
-        doc_str, end_line = get_document_with_lines(document_lines, start_line=45, max_characters=1000)
-        self.assertLess(end_line, len(document_lines))
-        
-        # Test with start_line at end of document
-        doc_str, end_line = get_document_with_lines(document_lines, start_line=49, max_characters=1000)
-        self.assertEqual(end_line, len(document_lines) - 1)
-
-    def test_section_continuity(self):
-        """Test that sections remain continuous across multiple processing iterations"""
-        # Create a document that will require multiple iterations
-        document_lines = [{"content": f"Line {i}"} for i in range(200)]
-        
-        all_sections = []
-        start_line = 0
-        max_iterations = 3
-        
-        for _ in range(max_iterations):
-            doc_str, end_line = get_document_with_lines(document_lines, start_line, max_characters=100)
-            
-            # Create some test sections
-            new_sections = [
-                DocumentSection(title=f"Section {i}", start_index=start_line + i*10)
-                for i in range(3)
-            ]
-            
-            # Validate and fix sections
-            fixed_sections = validate_and_fix_sections(new_sections, len(document_lines))
-            
-            if not fixed_sections:
-                start_line = end_line + 1
-                continue
-            
-            all_sections.extend(fixed_sections)
-            
-            if end_line >= len(document_lines) - 1:
-                break
-            
-            if len(fixed_sections) > 1:
-                # Test the overlap handling logic
-                start_line = all_sections[-1].start_index
-                all_sections.pop()
-            else:
-                start_line = end_line + 1
-            
-            self.assertLess(start_line, len(document_lines), 
-                           "start_line exceeded document length")
-        
-        # Convert to final sections
-        final_sections = get_sections_text(all_sections, document_lines)
-        
-        # Verify final sections
-        for section in final_sections:
-            self.assertLess(section["end"], len(document_lines), 
-                           f"Section end {section['end']} exceeds document length {len(document_lines)}")
-            self.assertGreaterEqual(section["start"], 0, 
-                                  f"Section start {section['start']} is negative")
+        # Should have a single section containing the entire document
+        self.assertEqual(len(sections), 1)
+        self.assertEqual(sections[0]["start"], 0)
+        self.assertEqual(sections[0]["end"], len(document_lines) - 1)
 
 if __name__ == "__main__":
     unittest.main()
